+++ oarepo_ui/theme/assets/semantic-ui/js/oarepo_ui/forms/components/EDTFDatePickerField/index.js
@@ -1,2 +1,2 @@
 export * from "./EDTFDateRangePickerField";
-export * from "./EDTFSingleDatePicker";
+export * from "./EDTFSingleDatePickerField";
